添加相机卡片之后再添备忘录卡片，桌面会奔溃

+++ frameworks/core/components/form/resource/form_request_data.h
@@ -21,7 +21,7 @@
 namespace OHOS::Ace {
 
 struct RequestFormInfo {
-    int32_t id = 0;
+    int64_t id = 0;
     std::string cardName;
     std::string bundleName;
     std::string abilityName;
